requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
jupyter>=1.0.0
ipykernel>=6.25.0
flask>=2.3.0
opencv-python>=4.8.0
optuna>=3.0.0
shap>=0.42.0

